+++ packages/logger/CHANGELOG.md
@@ -21,6 +21,28 @@
   * devDependencies
     * @libp2p/peer-id bumped from ^4.0.2 to ^4.0.3
 
+## [5.1.8](https://github.com/libp2p/js-libp2p/compare/logger-v5.1.7...logger-v5.1.8) (2025-02-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.1 to ^2.5.0
+  * devDependencies
+    * @libp2p/peer-id bumped from ^5.0.11 to ^5.0.12
+
+## [5.1.7](https://github.com/libp2p/js-libp2p/compare/logger-v5.1.6...logger-v5.1.7) (2025-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.0 to ^2.4.1
+  * devDependencies
+    * @libp2p/peer-id bumped from ^5.0.10 to ^5.0.11
+
 ## [5.1.6](https://github.com/libp2p/js-libp2p/compare/logger-v5.1.5...logger-v5.1.6) (2025-01-07)