+++ packages/metrics-devtools/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## [1.2.3](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.2.2...devtools-metrics-v1.2.3) (2025-02-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.1 to ^2.5.0
+    * @libp2p/interface-internal bumped from ^2.2.4 to ^2.3.0
+    * @libp2p/logger bumped from ^5.1.7 to ^5.1.8
+    * @libp2p/peer-id bumped from ^5.0.11 to ^5.0.12
+    * @libp2p/simple-metrics bumped from ^1.3.1 to ^1.3.2
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.10 to ^5.0.11
+
+## [1.2.2](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.2.1...devtools-metrics-v1.2.2) (2025-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^2.2.3 to ^2.2.4
+
+## [1.2.1](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.2.0...devtools-metrics-v1.2.1) (2025-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.0 to ^2.4.1
+    * @libp2p/interface-internal bumped from ^2.2.2 to ^2.2.3
+    * @libp2p/logger bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/peer-id bumped from ^5.0.10 to ^5.0.11
+    * @libp2p/simple-metrics bumped from ^1.3.0 to ^1.3.1
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.9 to ^5.0.10
+
 ## [1.2.0](https://github.com/libp2p/js-libp2p/compare/devtools-metrics-v1.1.12...devtools-metrics-v1.2.0) (2025-01-07)