+++ packages/metrics-opentelemetry/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
+## [1.0.3](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.2...opentelemetry-metrics-v1.0.3) (2025-02-10)
+
+
+### Bug Fixes
+
+* include platform in user agent ([#2942](https://github.com/libp2p/js-libp2p/issues/2942)) ([96f14e4](https://github.com/libp2p/js-libp2p/commit/96f14e429eac84d02504c4b97f183511c8af2add))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.1 to ^2.5.0
+    * @libp2p/utils bumped from ^6.5.0 to ^6.5.1
+
+## [1.0.2](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.1...opentelemetry-metrics-v1.0.2) (2025-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/utils bumped from ^6.4.0 to ^6.5.0
+
+## [1.0.1](https://github.com/libp2p/js-libp2p/compare/opentelemetry-metrics-v1.0.0...opentelemetry-metrics-v1.0.1) (2025-02-03)
+
+
+### Bug Fixes
+
+* use isPromise/etc function from utils ([#2925](https://github.com/libp2p/js-libp2p/issues/2925)) ([a32fbeb](https://github.com/libp2p/js-libp2p/commit/a32fbeb1dbf1ffbb59445f56ba011af4123ab085))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.0 to ^2.4.1
+    * @libp2p/utils bumped from ^6.3.1 to ^6.4.0
+
 ## 1.0.0 (2025-01-07)