+++ packages/metrics-prometheus/CHANGELOG.md
@@ -81,6 +81,52 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [4.3.4](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.3...prometheus-metrics-v4.3.4) (2025-02-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.2.6 to ^6.2.7
+
+## [4.3.3](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.2...prometheus-metrics-v4.3.3) (2025-02-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.1 to ^2.5.0
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.10 to ^5.0.11
+    * @libp2p/interface-compliance-tests bumped from ^6.2.5 to ^6.2.6
+    * @libp2p/logger bumped from ^5.1.7 to ^5.1.8
+    * @libp2p/peer-id bumped from ^5.0.11 to ^5.0.12
+
+## [4.3.2](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.1...prometheus-metrics-v4.3.2) (2025-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.2.4 to ^6.2.5
+
+## [4.3.1](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.0...prometheus-metrics-v4.3.1) (2025-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.0 to ^2.4.1
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.9 to ^5.0.10
+    * @libp2p/interface-compliance-tests bumped from ^6.2.3 to ^6.2.4
+    * @libp2p/logger bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/peer-id bumped from ^5.0.10 to ^5.0.11
+
 ## [4.3.0](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.2.10...prometheus-metrics-v4.3.0) (2025-01-07)