+++ packages/transport-memory/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## [1.0.5](https://github.com/libp2p/js-libp2p/compare/memory-v1.0.4...memory-v1.0.5) (2025-02-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.1 to ^2.5.0
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.7 to ^5.1.8
+    * @libp2p/peer-id bumped from ^5.0.11 to ^5.0.12
+
+## [1.0.4](https://github.com/libp2p/js-libp2p/compare/memory-v1.0.3...memory-v1.0.4) (2025-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.4.0 to ^2.4.1
+  * devDependencies
+    * @libp2p/logger bumped from ^5.1.6 to ^5.1.7
+    * @libp2p/peer-id bumped from ^5.0.10 to ^5.0.11
+
 ## [1.0.3](https://github.com/libp2p/js-libp2p/compare/memory-v1.0.2...memory-v1.0.3) (2025-01-07)